# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RIA laravel API",
        "version": "1.0"
    },
    "paths": {
        "/api/": {
            "get": {
                "summary": "Listing products",
                "operationId": "d60f6532d730ced60e88ef335d823570",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "products": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "title": {
                                                    "type": "string",
                                                    "maxLength": 50,
                                                    "minLength": 2
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "maxLength": 200,
                                                    "minLength": 10
                                                },
                                                "price": {
                                                    "type": "number",
                                                    "format": "float"
                                                },
                                                "created_at": {
                                                    "type": "time"
                                                },
                                                "updated_at": {
                                                    "type": "time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "id": 1,
                                    "title": "Самса",
                                    "description": "Очень вкусная, с курицей",
                                    "price": "219.99",
                                    "created_at": "2023-07-06T08:27:30.000000Z",
                                    "updated_at": "2023-07-06T09:45:07.000000Z"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/title={title}": {
            "get": {
                "summary": "Listing a product by title",
                "operationId": "0339dccec6ac15728febcd76f4fb9657",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "products": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "title": {
                                                    "type": "string",
                                                    "maxLength": 50,
                                                    "minLength": 2
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "maxLength": 200,
                                                    "minLength": 10
                                                },
                                                "price": {
                                                    "type": "number",
                                                    "format": "float"
                                                },
                                                "created_at": {
                                                    "type": "time"
                                                },
                                                "updated_at": {
                                                    "type": "time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "id": 1,
                                    "title": "Самса",
                                    "description": "Очень вкусная, с курицей",
                                    "price": "219.99",
                                    "created_at": "2023-07-06T08:27:30.000000Z",
                                    "updated_at": "2023-07-06T09:45:07.000000Z"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/{id}": {
            "get": {
                "summary": "Listing a product by id",
                "operationId": "e717cfab0bacaf599213f873b568de73",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "products": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "title": {
                                                    "type": "string",
                                                    "maxLength": 50,
                                                    "minLength": 2
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "maxLength": 200,
                                                    "minLength": 10
                                                },
                                                "price": {
                                                    "type": "number",
                                                    "format": "float"
                                                },
                                                "created_at": {
                                                    "type": "time"
                                                },
                                                "updated_at": {
                                                    "type": "time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "id": 1,
                                    "title": "Самса",
                                    "description": "Очень вкусная, с курицей",
                                    "price": "219.99",
                                    "created_at": "2023-07-06T08:27:30.000000Z",
                                    "updated_at": "2023-07-06T09:45:07.000000Z"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/create": {
            "post": {
                "summary": "Сreating a product according to the required parameters 'title', 'description', 'price'",
                "operationId": "cae2d9aa19f411f19468ab75e26a142a",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "products": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "title": {
                                                    "type": "string",
                                                    "maxLength": 50,
                                                    "minLength": 2
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "maxLength": 200,
                                                    "minLength": 10
                                                },
                                                "price": {
                                                    "type": "number",
                                                    "format": "float"
                                                },
                                                "created_at": {
                                                    "type": "time"
                                                },
                                                "updated_at": {
                                                    "type": "time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "id": 1,
                                    "title": "Самса",
                                    "description": "Очень вкусная, с курицей",
                                    "price": "219.99",
                                    "created_at": "2023-07-06T08:27:30.000000Z",
                                    "updated_at": "2023-07-06T09:45:07.000000Z"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/{id}/edit": {
            "patch": {
                "summary": "Product update by parameters 'title', 'description', 'price'",
                "operationId": "0a33468b04a0b2a543b542e791dcf70c",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "products": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "title": {
                                                    "type": "string",
                                                    "maxLength": 50,
                                                    "minLength": 2
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "maxLength": 200,
                                                    "minLength": 10
                                                },
                                                "price": {
                                                    "type": "number",
                                                    "format": "float"
                                                },
                                                "created_at": {
                                                    "type": "time"
                                                },
                                                "updated_at": {
                                                    "type": "time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "id": 1,
                                    "title": "Самса",
                                    "description": "Очень вкусная, с курицей",
                                    "price": "219.99",
                                    "created_at": "2023-07-06T08:27:30.000000Z",
                                    "updated_at": "2023-07-06T09:45:07.000000Z"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/{id}/delete": {
            "delete": {
                "summary": "Delete product by id",
                "operationId": "1e84495e8216bc744ebec2005b3e03d8",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "products": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "OK"
                    }
                }
            }
        }
    }
}